Moved server-side GitRepo code into its own module
