+++ settings.gradle
@@ -143,6 +143,9 @@ include 'stroom-explorer:stroom-explorer-impl-db-jooq'
 include 'stroom-feed:stroom-feed-api'
 include 'stroom-feed:stroom-feed-impl'
 
+// GitRepo
+include 'stroom-gitrepo'
+
 // Hadoop
 include 'stroom-hadoop'
 
@@ -294,3 +297,5 @@ include 'stroom-ssl-test'
 include 'stroom-view:stroom-view-api'
 include 'stroom-index:stroom-index-api'
 include 'stroom-cache:stroom-cache-service-impl'
+include 'stroom-gitrepo:stroom-gitrepo-impl'
+include 'stroom-gitrepo:stroom-gitrepo-api'
