+++ stroom-app/build.gradle
@@ -114,6 +114,7 @@ dependencies {
     implementation project(':stroom-explorer:stroom-explorer-impl-db-jooq')
     implementation project(':stroom-feed:stroom-feed-api')
     implementation project(':stroom-feed:stroom-feed-impl')
+    implementation project(':stroom-gitrepo:stroom-gitrepo-impl')
     implementation project(':stroom-importexport:stroom-importexport-api')
     implementation project(':stroom-importexport:stroom-importexport-impl')
     implementation project(':stroom-index:stroom-index-api')
