+++ stroom-gitrepo/.gitignore
@@ -1,0 +1,10 @@
+/.classpath
+/.externalToolBuilders/
+/gwt-DevMode/
+/gwt-SuperDevMode/
+/gwt-unitCache/
+/.project
+/.settings/
+/target/
+/build/
+/bin/
