+++ stroom-gitrepo/build.gradle
@@ -1,0 +1,1 @@
+ext.moduleName = 'stroom.gitrepo'
