+++ stroom-gitrepo/stroom-gitrepo-api/build.gradle
@@ -1,0 +1,11 @@
+ext.moduleName = 'stroom.gitrepo.api'
+
+dependencies {
+    implementation project(':stroom-core-shared')
+    implementation project(':stroom-docref')
+    implementation project(':stroom-docstore:stroom-docstore-api')
+    implementation project(':stroom-explorer:stroom-explorer-api')
+    implementation project(':stroom-importexport:stroom-importexport-api')
+    implementation project(':stroom-util-shared')
+    implementation project(':stroom-util-shared')
+}
