+++ stroom-gitrepo/stroom-gitrepo-api/src/main/java/stroom/gitrepo/api/GitRepoStore.java
@@ -15,7 +15,7 @@
  *
  */
 
-package stroom.dashboard.impl.gitrepo;
+package stroom.gitrepo.api;
 
 import stroom.docref.DocRef;
 import stroom.docstore.api.DocumentStore;
