+++ stroom-gitrepo/stroom-gitrepo-impl/build.gradle
@@ -1,0 +1,17 @@
+ext.moduleName = 'stroom.gitrepo.impl'
+
+dependencies {
+    implementation project(':stroom-core-shared')
+    implementation project(':stroom-docref')
+    implementation project(':stroom-docstore:stroom-docstore-api')
+    implementation project(':stroom-event-logging:stroom-event-logging-api')
+    implementation project(':stroom-event-logging-rs:stroom-event-logging-rs-api')
+    implementation project(':stroom-explorer:stroom-explorer-api')
+    implementation project(':stroom-gitrepo:stroom-gitrepo-api')
+    implementation project(':stroom-importexport:stroom-importexport-api')
+    implementation project(':stroom-util')
+    implementation project(':stroom-util-shared')
+    implementation libs.guice
+    implementation libs.jakarta.inject
+    implementation libs.restygwt
+}
