+++ stroom-gitrepo/stroom-gitrepo-impl/src/main/java/stroom/gitrepo/impl/GitRepoResourceImpl.java
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-package stroom.dashboard.impl.gitrepo;
+package stroom.gitrepo.impl;
 
 import stroom.docref.DocRef;
 import stroom.docstore.api.DocumentResourceHelper;
 import stroom.event.logging.rs.api.AutoLogged;
+import stroom.gitrepo.api.GitRepoStore;
 import stroom.gitrepo.shared.GitRepoDoc;
 import stroom.gitrepo.shared.GitRepoResource;
 import stroom.util.shared.EntityServiceException;
